Travis CI: download WordPoints' PHPCS config
Previously I had added a linen that copied it from the download of
WordPoints. However, we only download WordPoints when we are running
the PHPUnit pass, so it doesn’t exist for the code sniff pass. Now we
download just the PHPCS XML config file from WordPoints when setting up
for the code sniff pass.

See #6