Update EDD SL API's get_module_info method
First, we change the `$module` method to be `$module_id`, since this is
the only piece of module data that we use. The complete module data was
originally being required to use in a request to the remote channel if
no cached data was found. However, it was decided not to do this.

We also change the `$info` parameter to be `$key` instead, for better
parity with a `set_module_information()` method (to be introduced).

See #19