Make sure that `$response['sections']` is a string
The EDD SL Free API will respond with this as an array. EDD SL
serializes the array into a string in the response, and so we need to
pass it through `maybe_unserialize()`. Of course, for the EDD SL Free
API, when we already have an array, we don’t want to attempt to pass it
through `preg_match()` or `maybe_unserialize()`. So we need to make
sure that what we have is a string before we do that.

See #19